work in progress change. Includes reimplementation of the sequence buffer to handle edge cases

+++ src/config.rs
@@ -23,7 +23,7 @@ pub struct Config {
     /// This is the maximum size of each fragment. It defaults to `1450` bytes, due to the default MTU on most network devices being `1500`.
     pub fragment_size: u16,
     /// Value which can specify the size of the buffer that queues up fragments ready to be reassembled once all fragments have arrived.```
-    pub fragment_reassembly_buffer_size: usize,
+    pub fragment_reassembly_buffer_size: u16,
     /// Value that specifies the size of the buffer the UDP data will be read into. Defaults to `1450` bytes.
     pub receive_buffer_max_size: usize,
     /// Value which can specify the factor which will smooth out network jitter.
